296. Validações parte 1 - Controle de fluxosE

+++ src/app/Http/Controllers/MarcaController.php
@@ -8,6 +8,11 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    public function __construct(Marca $marca)
+    {
+        $this->marca = $marca;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -44,11 +49,17 @@ class MarcaController extends Controller
      * Display the specified resource.
      *
      * @param \App\Models\Marca $marca
-     * @return Marca
+     * @return string[]
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
-        return $marca;
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return ['erro' => 'Marca não existe'];
+        } else {
+            return $marca;
+        }
     }
 
     /**
@@ -67,12 +78,17 @@ class MarcaController extends Controller
      *
      * @param \Illuminate\Http\Request $request
      * @param \App\Models\Marca $marca
-     * @return Marca
+     * @return Marca|string[]
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
+        $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Atualização não realizada, marca não existe'];
+        }
         $marca->update($request->all());
         return $marca;
+
     }
 
     /**
@@ -81,8 +97,12 @@ class MarcaController extends Controller
      * @param \App\Models\Marca $marca
      * @return string[]
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Deleção não realizada, marca não existe'];
+        }
         $marca->delete();
         return ['msg' => 'Marca deletada com sucesso.'];
     }
